requirements.txt
beautifulsoup4
langchain
openai
pandas
pydantic
requests
streamlit
openpyxl
googlesearch-python
# Optional:
